+++ src/components/app/actions.js
@@ -1,11 +1,11 @@
-import initialState from '../../initialState';
+import {INITIAL_STATE} from './reducer';
 
 export default {
   changeCity: (city) => ({type: `CHANGE_CITY`, payload: city}),
 
   getOffers: (city) => ({
     type: `GET_OFFERS`,
-    payload: initialState
+    payload: INITIAL_STATE
       .offers
       .filter((item) => item.city === city)
   }),
