+++ src/components/app/actions.test.js
@@ -1,5 +1,5 @@
-import appActions from './app.actions';
-import initialState from '../../initialState';
+import appActions from './actions';
+import {INITIAL_STATE} from './reducer';
 
 describe(`App actions creator work correctly`, () => {
   it(`App actions creator return change city type when city was changed`, () => {
@@ -9,7 +9,7 @@ describe(`App actions creator work correctly`, () => {
   it(`App actions creator return get offers type with offers in city when city was changed`, () => {
     expect(appActions.getOffers(`Amsterdam`)).toEqual({
       type: `GET_OFFERS`,
-      payload: initialState
+      payload: INITIAL_STATE
         .offers
         .filter((item) => item.city === `Amsterdam`)
     });
