+++ src/index.js
@@ -1,4 +1,4 @@
-import createStore from './store/createStore';
+import store from './store';
 import {Provider} from 'react-redux';
 import ReactDOM from 'react-dom';
 import App from 'Components/app/app.connect';
@@ -6,7 +6,6 @@ import rentMockList from './mocks/offers';
 import {getActualCities} from './utils';
 
 const init = () => {
-  const store = createStore();
   const offers = store
     .getState()
     .app
