--- src/initialState.js
@@ -1,7 +1,0 @@
-import mockOffers from './mocks/offers';
-
-export default {
-  city: `Amsterdam`,
-  sortType: `Popular`,
-  offers: mockOffers
-};
