--- src/store/createStore.js
@@ -1,4 +1,0 @@
-import {createStore} from 'redux';
-import rootReducer from '../rootReducer';
-
-export default () => createStore(rootReducer);
