+++ src/store/reducer.js
@@ -1,5 +1,5 @@
 import {combineReducers} from 'redux';
-import appReducer from './components/app/app.reducer';
+import appReducer from '../components/app/reducer';
 
 const rootReducer = combineReducers({app: appReducer});
