+++ src/store/store.js
@@ -1,0 +1,4 @@
+import {createStore} from 'redux';
+import rootReducer from './reducer';
+
+export default createStore(rootReducer);
